#174 adding changes for cricinfo crawler
