+++ src/main/java/com/garv/satta/fantasy/external/service/cricinfo/parser/BowlingParser.java
@@ -1,0 +1,7 @@
+package com.garv.satta.fantasy.external.service.cricinfo.parser;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class BowlingParser {
+}
