+++ src/main/resources/application.properties
@@ -42,4 +42,4 @@ spring.cloud.gcp.sql.database-name=satta
 
 fantasy.cricapi.url = https://cricapi.com/api/
 fantasy.cricapi.key =  key
-fixedDelay.in.milliseconds = 300000
+fixedDelay.in.milliseconds = 30000
